requirements.txt
streamlit
requests
pandas
numpy
xgboost
scikit-learn
python-dotenv
plotly
